backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
plotly==5.18.0
scipy==1.11.4